changed key back to int from long

+++ Server/Tasks/Application/Models/GetAllProjectsModel.cs
@@ -8,6 +8,6 @@ namespace Core.Application.Models
     public class GetAllProjectsModel
     {
         [Required]
-        public long UserID;
+        public int UserID;
     }
 }
